[refactor/TK-226]: controller 위치 변경
