--- src/main/java/withbeetravel/controller/notification/NotificationController.java
@@ -1,36 +1,0 @@
-package withbeetravel.controller.notification;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestHeader;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
-import withbeetravel.security.UserAuthorizationUtil;
-import withbeetravel.service.notification.NotificationService;
-
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-
-@RestController
-@Slf4j
-@RequiredArgsConstructor
-@RequestMapping("/api/notifications")
-public class NotificationController {
-
-    private final NotificationService notificationService;
-    // 모든 Emitters를 저장하는 ConcurrentHashMap
-    public static Map<Long, SseEmitter> sseEmitters = new ConcurrentHashMap<>();
-
-    /**
-     * streamNotifications : 사용자의 요청을 받아 SSE 구독하도록 연결
-     * - SseEmitter를 반환
-     */
-    @GetMapping(value = "/stream", produces = "text/event-stream")
-    public SseEmitter streamNotifications(@RequestHeader(value = "Last-Event-ID", required = false, defaultValue = "")
-                                              String lastEventId) {
-        Long userId = UserAuthorizationUtil.getLoginUserId();
-        return notificationService.subscribe(userId, lastEventId);
-    }
-}
